Refactor CaseListService into its own submodule
